+++ lib/src/date_helper.i18n.dart
@@ -7,164 +7,164 @@ extension Localization on String {
   String get i18n => localize(this, _t);
 }
 
-final _t = Translations('en_us') +
+final _t = Translations('en') +
     {
-      'en_us': 'MM/DD/YYYY',
-      'de_de': 'YYYY-MM-DD',
+      'en': 'MM/DD/YYYY',
+      'de': 'YYYY-MM-DD',
     } +
     // --- Days in week - long
     {
-      'en_us': 'Monday',
-      'de_de': 'Montag',
+      'en': 'Monday',
+      'de': 'Montag',
     } +
     {
-      'en_us': 'Tuesday',
-      'de_de': 'Dienstag',
+      'en': 'Tuesday',
+      'de': 'Dienstag',
     } +
     {
-      'en_us': 'Wednesday',
-      'de_de': 'Mittwoch',
+      'en': 'Wednesday',
+      'de': 'Mittwoch',
     } +
     {
-      'en_us': 'Thursday',
-      'de_de': 'Donnerstag',
+      'en': 'Thursday',
+      'de': 'Donnerstag',
     } +
     {
-      'en_us': 'Friday',
-      'de_de': 'Freitag',
+      'en': 'Friday',
+      'de': 'Freitag',
     } +
     {
-      'en_us': 'Saturday',
-      'de_de': 'Samstag',
+      'en': 'Saturday',
+      'de': 'Samstag',
     } +
     {
-      'en_us': 'Sunday',
-      'de_de': 'Sonntag',
+      'en': 'Sunday',
+      'de': 'Sonntag',
     } +
     // --- Days in week - short
     {
-      'en_us': 'Mo',
-      'de_de': 'Mo',
+      'en': 'Mo',
+      'de': 'Mo',
     } +
     {
-      'en_us': 'Tu',
-      'de_de': 'Di',
+      'en': 'Tu',
+      'de': 'Di',
     } +
     {
-      'en_us': 'We',
-      'de_de': 'Mi',
+      'en': 'We',
+      'de': 'Mi',
     } +
     {
-      'en_us': 'Th',
-      'de_de': 'Do',
+      'en': 'Th',
+      'de': 'Do',
     } +
     {
-      'en_us': 'Fr',
-      'de_de': 'Fr',
+      'en': 'Fr',
+      'de': 'Fr',
     } +
     {
-      'en_us': 'Sa',
-      'de_de': 'Sa',
+      'en': 'Sa',
+      'de': 'Sa',
     } +
     {
-      'en_us': 'Su',
-      'de_de': 'So',
+      'en': 'Su',
+      'de': 'So',
     } +
     // --- Months in year - long
     {
-      'en_us': 'January',
-      'de_de': 'Januar',
+      'en': 'January',
+      'de': 'Januar',
     } +
     {
-      'en_us': 'February',
-      'de_de': 'Februar',
+      'en': 'February',
+      'de': 'Februar',
     } +
     {
-      'en_us': 'March',
-      'de_de': 'März',
+      'en': 'March',
+      'de': 'März',
     } +
     {
-      'en_us': 'April',
-      'de_de': 'April',
+      'en': 'April',
+      'de': 'April',
     } +
     {
-      'en_us': 'May',
-      'de_de': 'Mai',
+      'en': 'May',
+      'de': 'Mai',
     } +
     {
-      'en_us': 'June',
-      'de_de': 'Juni',
+      'en': 'June',
+      'de': 'Juni',
     } +
     {
-      'en_us': 'July',
-      'de_de': 'Juli',
+      'en': 'July',
+      'de': 'Juli',
     } +
     {
-      'en_us': 'August',
-      'de_de': 'August',
+      'en': 'August',
+      'de': 'August',
     } +
     {
-      'en_us': 'September',
-      'de_de': 'September',
+      'en': 'September',
+      'de': 'September',
     } +
     {
-      'en_us': 'October',
-      'de_de': 'Oktober',
+      'en': 'October',
+      'de': 'Oktober',
     } +
     {
-      'en_us': 'November',
-      'de_de': 'November',
+      'en': 'November',
+      'de': 'November',
     } +
     {
-      'en_us': 'December',
-      'de_de': 'Dezember',
+      'en': 'December',
+      'de': 'Dezember',
     } +
     // --- Months in year - short
     {
-      'en_us': 'Jan',
-      'de_de': 'Jan',
+      'en': 'Jan',
+      'de': 'Jan',
     } +
     {
-      'en_us': 'Feb',
-      'de_de': 'Feb',
+      'en': 'Feb',
+      'de': 'Feb',
     } +
     {
-      'en_us': 'Mar',
-      'de_de': 'Mär',
+      'en': 'Mar',
+      'de': 'Mär',
     } +
     {
-      'en_us': 'Apr',
-      'de_de': 'Apr',
+      'en': 'Apr',
+      'de': 'Apr',
     } +
     {
-      'en_us': 'May',
-      'de_de': 'Mai',
+      'en': 'May',
+      'de': 'Mai',
     } +
     {
-      'en_us': 'Jun',
-      'de_de': 'Jun',
+      'en': 'Jun',
+      'de': 'Jun',
     } +
     {
-      'en_us': 'Jul',
-      'de_de': 'Jul',
+      'en': 'Jul',
+      'de': 'Jul',
     } +
     {
-      'en_us': 'Aug',
-      'de_de': 'Aug',
+      'en': 'Aug',
+      'de': 'Aug',
     } +
     {
-      'en_us': 'Sep',
-      'de_de': 'Sep',
+      'en': 'Sep',
+      'de': 'Sep',
     } +
     {
-      'en_us': 'Oct',
-      'de_de': 'Okt',
+      'en': 'Oct',
+      'de': 'Okt',
     } +
     {
-      'en_us': 'Nov',
-      'de_de': 'Nov',
+      'en': 'Nov',
+      'de': 'Nov',
     } +
     {
-      'en_us': 'Dec',
-      'de_de': 'Dez',
+      'en': 'Dec',
+      'de': 'Dez',
     };